+++ package.json
@@ -1,63 +1,78 @@
 {
-    "name": "my-app",
-    "version": "0.2.0",
-    "private": true,
-    "dependencies": {
-        "@types/react-qr-reader": "^2.1.4",
-        "@vkontakte/icons": "^1.101.0",
-        "@vkontakte/vk-bridge": "^2.4.2",
-        "@vkontakte/vkjs": "^0.20.0",
-        "@vkontakte/vkui": "^4.13.0",
-        "axios": "^0.21.0",
-        "mobx": "^6.3.2",
-        "mobx-react-lite": "^3.2.0",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-hook-form": "^7.7.1",
-        "react-lottie-player": "^1.3.3",
-        "react-qr-reader": "^2.2.1",
-        "react-query": "^3.16.0",
-        "react-scripts": "4.0.3",
-        "styled-components": "^5.3.0",
-        "typescript": "~4.3.2"
-    },
-    "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject"
-    },
-    "eslintConfig": {
-        "extends": "react-app"
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
-    },
-    "devDependencies": {
-        "@testing-library/jest-dom": "^5.11.9",
-        "@testing-library/react": "^11.2.3",
-        "@testing-library/user-event": "^12.6.0",
-        "@types/jest": "^26.0.20",
-        "@types/node": "^14.14.21",
-        "@types/react": "^17.0.0",
-        "@types/react-dom": "^17.0.0",
-        "@types/styled-components": "^5.1.2",
-        "@typescript-eslint/eslint-plugin": "^4.13.0",
-        "@typescript-eslint/parser": "^4.13.0",
-        "eslint-config-standard": "^16.0.2",
-        "eslint-plugin-import": "^2.22.0",
-        "eslint-plugin-node": "^11.1.0",
-        "eslint-plugin-promise": "^4.2.1",
-        "eslint-plugin-react": "^7.20.5",
-        "eslint-plugin-standard": "^5.0.0"
-    }
+  "name": "my-app",
+  "version": "0.2.0",
+  "private": true,
+  "dependencies": {
+    "@vkontakte/icons": "^1.125.0",
+    "@vkontakte/vk-bridge": "^2.4.8",
+    "@vkontakte/vkui": "^4.14.0",
+    "axios": "^0.21.0",
+    "history": "^5.0.1",
+    "mobx": "^6.3.2",
+    "mobx-react-lite": "^3.2.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-hook-form": "^7.7.1",
+    "react-lottie-player": "^1.3.3",
+    "react-qr-reader": "^2.2.1",
+    "react-query": "^3.16.0",
+    "react-router5": "^8.0.1",
+    "react-scripts": "^4.0.3",
+    "router5": "^8.0.1",
+    "router5-plugin-browser": "^8.0.1",
+    "router5-plugin-logger": "^8.0.1",
+    "styled-components": "^5.3.0",
+    "typescript": "~4.3.2"
+  },
+  "sideEffects": false,
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "prepare": "husky install",
+    "pre-commit": "lint-staged"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "lint-staged": {
+    "*.(ts|tsx)": [
+      "./node_modules/.bin/eslint src --max-warnings=0",
+      "pretty-quick --staged"
+    ]
+  },
+  "devDependencies": {
+    "@types/node": "^14.14.21",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
+    "@types/react-qr-reader": "^2.1.4",
+    "@types/styled-components": "^5.1.2",
+    "@typescript-eslint/eslint-plugin": "^4.13.0",
+    "@typescript-eslint/parser": "^4.13.0",
+    "eslint": "^7.32.0",
+    "eslint-config-standard": "^16.0.2",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.20.5",
+    "eslint-plugin-standard": "^5.0.0",
+    "husky": "^7.0.2",
+    "lint-staged": "^11.1.2",
+    "prettier": "^2.3.2",
+    "pretty-quick": "^3.1.1",
+    "react-is": "^17.0.2"
+  }
 }
