+++ src/App.tsx
@@ -10,30 +10,30 @@ import { UsersAPI } from "./features/utils/requests/user-request";
 import { NoBoecAttached } from "./ui/molecules/NoBoecAttached";
 
 export const App: FC = observer(({ children }) => {
-    const { colorSchema, isInitialization, setUser, user } =
-        useContext(appStore);
-    useQuery({
-        queryKey: ["user-me"],
-        queryFn: () => UsersAPI.getMeData(),
-        retry: 1,
-        refetchInterval: 60000,
-        refetchOnWindowFocus: false,
-        enabled: !isInitialization,
-        onSuccess: setUser,
-    });
-    return (
-        <ConfigProvider scheme={colorSchema}>
-            <AdaptivityProvider>
-                <AppRoot>
-                    {isInitialization ? (
-                        <Intro />
-                    ) : user?.boec !== null ? (
-                        <AppLayout>{children}</AppLayout>
-                    ) : (
-                        <NoBoecAttached />
-                    )}
-                </AppRoot>
-            </AdaptivityProvider>
-        </ConfigProvider>
-    );
+  const { colorSchema, isInitialization, setUser, user } = useContext(appStore);
+
+  useQuery({
+    queryKey: ["user-me"],
+    queryFn: () => UsersAPI.getMeData(),
+    retry: 1,
+    refetchInterval: 60000,
+    refetchOnWindowFocus: false,
+    enabled: !isInitialization,
+    onSuccess: setUser,
+  });
+  return (
+    <ConfigProvider scheme={colorSchema}>
+      <AdaptivityProvider>
+        <AppRoot>
+          {isInitialization ? (
+            <Intro />
+          ) : user?.boec !== null ? (
+            <AppLayout>{children}</AppLayout>
+          ) : (
+            <NoBoecAttached />
+          )}
+        </AppRoot>
+      </AdaptivityProvider>
+    </ConfigProvider>
+  );
 });
