+++ src/AppLayout.tsx
@@ -1,103 +1,75 @@
-import { FC, useContext, useEffect } from "react";
+import { FC, useContext } from "react";
 import { observer } from "mobx-react-lite";
 import {
-    Epic,
-    usePlatform,
-    useAdaptivity,
-    ViewWidth,
-    VKCOM,
-    SplitCol,
-    SplitLayout,
-    PanelHeader,
-    Root,
-    SizeType,
+  usePlatform,
+  useAdaptivity,
+  ViewWidth,
+  VKCOM,
+  SplitCol,
+  SplitLayout,
+  PanelHeader,
 } from "@vkontakte/vkui";
 
-import { routerStore } from "./features/stores/router-store";
 import { DesktopMenu } from "./ui/molecules/DesktopMenu";
-import { MobileMenu } from "./ui/molecules/MobileMenu";
 import { Modals } from "./ui/organisms/Modals";
-import { toJS } from "mobx";
-import { BoecView } from "./features/boec/template/BoecView";
-import { BrigadesView } from "./features/brigades/template/BrigadesView";
-import { ElseView } from "./features/else/pages/ElseView";
-import { EventView } from "./features/event/template/EventView";
-import { ShtabView } from "./features/shtab/template/ShtabView";
-import { ProfileView } from "./features/profile/template/ProfileView";
-import { ScannerView } from "./features/scanner/template/ScannerView";
+import { routerStore } from "./features/stores/router-store";
+
+import { useRoute } from "react-router5";
 
 export const AppLayout: FC = observer(() => {
-    const {
-        setLastAndroidBackAction,
-        lastAndroidBackAction,
-        goBack,
-        activeStory,
-        activeView,
-        popout,
-        pageScrollPosition,
-    } = useContext(routerStore);
-    const router = useContext(routerStore);
-    console.log(toJS(router));
-    useEffect(() => {
-        const popListener = () => {
-            let timeNow = +new Date();
-            if (timeNow - lastAndroidBackAction > 500) {
-                setLastAndroidBackAction(timeNow);
-                goBack();
-            } else {
-                window.history.pushState(null, "");
-            }
-        };
-        window.addEventListener("popstate", popListener);
-        return () => {
-            window.removeEventListener("popstate", popListener);
-        };
-    }, [goBack, lastAndroidBackAction, setLastAndroidBackAction]);
+  const { popout } = useContext(routerStore);
 
-    useEffect(() => {
-        window.scroll(0, pageScrollPosition);
-    }, [pageScrollPosition]);
+  const platform = usePlatform();
+  const { viewWidth = 3 } = useAdaptivity();
+  const { route } = useRoute();
+  const isDesktop = viewWidth >= ViewWidth.SMALL_TABLET;
+  const hasHeader = platform !== VKCOM;
+  // const location = useMemo(() => route.name.split("."), [route.name]);
+  console.log(route);
 
-    const platform = usePlatform();
-    const { viewWidth = 3 } = useAdaptivity();
-
-    const isDesktop = viewWidth >= ViewWidth.SMALL_TABLET;
-    const hasHeader = platform !== VKCOM;
-
-    return (
-        <SplitLayout
-            modal={<Modals />}
-            popout={popout}
-            header={hasHeader && <PanelHeader separator={false} />}
-            style={{ justifyContent: "center" }}
-        >
-            <SplitCol
-                animate={!isDesktop}
-                spaced={isDesktop}
-                width={isDesktop ? "560px" : "100%"}
-                maxWidth={isDesktop ? "560px" : "100%"}
-            >
-                <Epic
+  return (
+    <SplitLayout
+      modal={<Modals />}
+      popout={popout}
+      header={hasHeader && <PanelHeader separator={false} />}
+      style={{ justifyContent: "center" }}
+    >
+      <SplitCol
+        animate={!isDesktop}
+        spaced={isDesktop}
+        width={isDesktop ? "560px" : "100%"}
+        maxWidth={isDesktop ? "560px" : "100%"}
+      >
+        {/* <Epic
                     activeStory={activeStory}
                     tabbar={!isDesktop && <MobileMenu />}
                     sizeY={SizeType.REGULAR}
                 >
-                    <Root id="else" activeView={activeView}>
-                        <ElseView id="else" />
-                        <EventView id="event" />
-                        <BrigadesView id="brigades" />
-                        <BoecView id="boec" />
-                        <ShtabView id="shtab" />
+                    <Root nav="else" activeView={activeView}>
+                        <View activePanel={activePanel} nav="base">
+                            <ElsePanel nav="base" />
+                        </View>
+                        <View activePanel={activePanel} nav="event">
+                            <EventListPanel nav="base" />
+                        </View>
                     </Root>
-                    <Root id="profile" activeView={activeView}>
-                        <ProfileView id="profile" />
+                    <Root nav="scanner" activeView={activeView}>
+                        <View nav="base" activePanel={activePanel}>
+                            <ScannerPanel nav="base" />
+                            <ScanPanel nav="scan" />
+                        </View>
                     </Root>
-                    <Root id="scanner" activeView={activeView}>
-                        <ScannerView id="scanner" />
+                    <Root nav="profile" activeView={activeView}>
+                        <View activePanel={activePanel} nav="base">
+                            <ProfilePanel nav="base" />
+                        </View>
+                        <View activePanel={activePanel} nav="notifications">
+                            <NotificationsPanel nav="base" />
+                        </View>
                     </Root>
-                </Epic>
-            </SplitCol>
-            {isDesktop && <DesktopMenu />}
-        </SplitLayout>
-    );
+                </Epic> */}
+      </SplitCol>
+      {isDesktop && <DesktopMenu />}
+    </SplitLayout>
+  );
 });
