+++ src/ui/molecules/DesktopMenu/index.tsx
@@ -1,43 +1,35 @@
-import { Icon28Newsfeed, Icon28UserSquareOutline } from "@vkontakte/icons";
 import {
-    Badge,
-    Cell,
-    Group,
-    Panel,
-    PanelHeader,
-    SplitCol,
-    usePlatform,
-    VKCOM,
+  Group,
+  Panel,
+  PanelHeader,
+  SplitCol,
+  usePlatform,
+  VKCOM,
 } from "@vkontakte/vkui";
 import { observer } from "mobx-react-lite";
-import { useContext } from "react";
-import { appStore } from "../../../features/stores/app-store";
-import { routerStore } from "../../../features/stores/router-store";
 
 export const DesktopMenu = observer(() => {
-    const platform = usePlatform();
+  const platform = usePlatform();
 
-    const hasHeader = platform !== VKCOM;
-    const { activeStory, setStory } = useContext(routerStore);
-    const { user } = useContext(appStore);
+  const hasHeader = platform !== VKCOM;
 
-    return (
-        <SplitCol fixed width="280px" maxWidth="280px">
-            <Panel>
-                {hasHeader && <PanelHeader />}
-                <Group>
-                    <Cell
+  return (
+    <SplitCol fixed width="280px" maxWidth="280px">
+      <Panel>
+        {hasHeader && <PanelHeader />}
+        <Group>
+          {/* <Cell
                         disabled={activeStory === "else"}
                         style={
                             activeStory === "else"
                                 ? {
                                       backgroundColor:
                                           "var(--button_secondary_background)",
-                                      borderRadius: 8,
+                                      borderRadius: 8
                                   }
                                 : {}
                         }
-                        onClick={() => setStory("else", "base", "else")}
+                        onClick={() => route.router.navigateToDefault()}
                         before={<Icon28Newsfeed />}
                         id="base"
                     >
@@ -50,11 +42,11 @@ export const DesktopMenu = observer(() => {
                                 ? {
                                       backgroundColor:
                                           "var(--button_secondary_background)",
-                                      borderRadius: 8,
+                                      borderRadius: 8
                                   }
                                 : {}
                         }
-                        onClick={() => setStory("profile", "base", "profile")}
+                        onClick={() => route.router.navigate("profile")}
                         before={<Icon28UserSquareOutline />}
                         id="profile"
                         badge={
@@ -67,9 +59,9 @@ export const DesktopMenu = observer(() => {
                         }
                     >
                         Профиль
-                    </Cell>
-                </Group>
-            </Panel>
-        </SplitCol>
-    );
+                    </Cell> */}
+        </Group>
+      </Panel>
+    </SplitCol>
+  );
 });
